editoast: move ModelV2 macro args parsing into a dedicated module
